requirements.txt
asgiref==3.11.1
Django==6.0.3
google-auth==2.48.0
gspread==6.2.1
gunicorn==25.1.0
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
